Fix ubiquitous 'github-code-scanning' PR message
We have to specify the 'push' target to avoid the message

| You have successfully added a new vcs-diff-lint configuration
| .github/workflows/python-diff-lint.yml:python-lint-job. As part of the
| setup process, we have scanned this repository and found no existing
| alerts. In the future, you will see all code scanning alerts on the
| repository Security tab.

... appearing in every single PR.  Without the 'push' trigger defined,
GitHub thinks that the linter action is being added in every single
pull-request.

Though, since the workflow is newly run for 'push' events, we have to do
some reasonable "no-op".  For the first implementation, we simply try to
generate an empty list of issues (therefore the container fix here).

For the vcs-diff-lint case, it doesn't make too much sense to do the
analysis on pushes.  At least not in the "basic" differential use-case
we have now.  We could e.g. do full-scan in the future if there was
demand for it.

So I'm fixing the example in README.md to define "push", too.  While on
it, move the 'if' statements down so the rules are slightly more
readable.

Related: redhat-plumbers-in-action/differential-shellcheck#215
Closes: #17
Closes: #16